requirements.txt
openai==0.26.5
ttkbootstrap==1.10.1
python-dotenv==1.0.0
